+++ upload/admin/language/russian/module/rees46.php
@@ -1,13 +1,13 @@
 <?php
 // Heading
-$_['heading_title']               = '<img src="view/image/rees46.png" width="20"> REES46';
+$_['heading_title']               = 'REES46';
 
 // Tab
-$_['tab_settings']                = 'Общее';
-$_['tab_xml']                     = 'Товары';
+$_['tab_general']                 = 'Общее';
+$_['tab_products']                = 'Товары';
 $_['tab_orders']                  = 'Заказы';
-$_['tab_subscribers']             = 'Покупатели';
-$_['tab_webpush']                 = 'Web Push';
+$_['tab_customers']               = 'Покупатели';
+$_['tab_webpush']                 = 'Веб-пуш';
 $_['tab_modules']                 = 'Рекомендации';
 $_['tab_help']                    = 'Помощь';
 
@@ -21,25 +21,15 @@ $_['text_module']                 = 'Модули';
 $_['text_success']                = 'Настройки модуля REES46 обновлены!';
 $_['text_processing_orders']      = 'Вы экспортировали в REES46 <b>%s</b> из <b>%s</b> отобранных заказов!';
 $_['text_success_orders']         = 'Вы успешно экспортировали в REES46 все <b>%s</b> отобранных заказов!';
-$_['text_processing_subscribers'] = 'Вы экспортировали в REES46 <b>%s</b> из <b>%s</b> выбранных покупателей!';
-$_['text_success_subscribers']    = 'Вы успешно экспортировали в REES46 все <b>%s</b> выбранных покупателей!';
+$_['text_processing_customers']   = 'Вы экспортировали в REES46 <b>%s</b> из <b>%s</b> выбранных покупателей!';
+$_['text_success_customers']      = 'Вы успешно экспортировали в REES46 все <b>%s</b> выбранных покупателей!';
 $_['text_error_export']           = 'Ошибка: Нет данных для экспорта!';
 $_['text_success_check']          = 'Файл <b>%s</b> успешно загружен!';
 $_['text_error_check']            = 'Ошибка: Вам необходимо загрузить файл <b>%s</b>!';
-$_['text_success_shop']           = 'Информация о магазине успешно записана в XML файл!';
-$_['text_error_shop']             = 'Ошибка: Информация о магазине не была записана в XML файл!';
-$_['text_success_currencies']     = 'Валюты успешно записаны в XML файл!';
-$_['text_error_currencies']       = 'Ошибка: Валюты не были записаны в XML файл!';
-$_['text_error_currency_code']    = 'Ошибка: Недопустимый код основной валюты!';
-$_['text_success_categories']     = 'Категории успешно записаны в XML файл!';
-$_['text_error_categories']       = 'Ошибка: Категории не были записана в XML файл!';
-$_['text_success_offers']         = 'Товар <b>[%s]</b> успешно записан в XML файл!';
-$_['text_error_offers']           = 'Ошибка: Товар <b>[%s]</b> не был записан в XML файл!';
-$_['text_success_xml']            = 'Вы успешно сгенерировали XML файл!';
 $_['text_edit']                   = 'Настройки модуля';
 $_['text_tab_module']             = 'Модуль';
 $_['text_documentation']          = 'Документация';
-$_['text_documentation_url']      = 'http://docs.rees46.com/display/R46D/OpenCart+2';
+$_['text_documentation_url']      = 'http://docs.rees46.com/pages/viewpage.action?pageId=1835122';
 $_['text_type_interesting']       = 'Возможно, вам это понравится';
 $_['text_type_also_bought']       = 'С этим товаром покупают';
 $_['text_type_similar']           = 'Похожие товары';
@@ -58,29 +48,27 @@ $_['text_autocomplete']           = 'Автодополнение';
 $_['text_subscribers']            = 'Подписчики на новости';
 $_['text_customers']              = 'Все покупатели';
 $_['text_info_1']                 = 'Чтобы начать использовать этот модуль, пожалуйста, зарегистрируйтесь на <a href="https://rees46.com/opencart" target="_blank">rees46.com</a> и получите ключи API для этой формы.';
-$_['text_info_2']                 = 'Экспортируйте только один раз при первой настройке модуля. Будут экспортированы заказы за последние пол года с выбранными выше статусами. Пожалуйста, сохраните настройки перед экспортом!';
-$_['text_info_3']                 = 'Пожалуйста, сохраните настройки перед экспортом!';
-$_['text_info_4']                 = 'Сгенерируйте файл XML, после чего укажите ссылку на него в кабинете rees46. Пожалуйста, сохраните настройки перед экспортом!';
+$_['text_info_2']                 = 'Экспортируйте только один раз при первой настройке модуля. Будут экспортированы заказы за последние пол года с выбранными выше статусами. Пожалуйста, сохраните настройки перед экспортом.';
+$_['text_info_3']                 = 'Пожалуйста, сохраните настройки перед экспортом.';
+$_['text_info_4']                 = 'Для активации веб-пуш уведомлений необходимо: cделать так, чтобы сайт поддерживал протокол HTTPS, загрузить в корень сайта файлы manifest.json и push_sw.js. Нажмите "Проверить файлы".';
 
 // Entry
-$_['entry_shop_id']               = 'Shop id';
-$_['entry_secret_key']            = 'Secret key';
-$_['entry_status']                = 'Статус';
+$_['entry_store_key']             = 'Store Key';
+$_['entry_secret_key']            = 'Secret Key';
+$_['entry_status']                = 'Статус модуля';
 $_['entry_log']                   = 'Логирование';
 $_['entry_export_orders']         = 'Экспорт заказов';
 $_['entry_status_created']        = 'Статус нового заказа';
 $_['entry_status_completed']      = 'Статус завершенного заказа';
 $_['entry_status_cancelled']      = 'Статус отмененного заказа';
-$_['entry_export_subscribers']    = 'Экспорт подписчиков';
-$_['entry_export_type']           = 'Тип экспорта';
+$_['entry_export_customers']      = 'Экспорт покупателей';
+$_['entry_export_type']           = 'Тип покупателей';
 $_['entry_webpush_files']         = 'Необходимые файлы';
-$_['entry_xml_status']            = 'Статус XML файла';
-$_['entry_xml_currency']          = 'Валюта товаров в XML файле';
-$_['entry_xml_generate']          = 'Генерация XML файла';
-$_['entry_xml_url']               = 'Адрес XML файла';
-$_['entry_name']                  = 'Название модуля';
-$_['entry_title']                 = 'Заголовок';
-$_['entry_type']                  = 'Тип';
+$_['entry_xml_currency']          = 'Валюта';
+$_['entry_xml_cron']              = 'Ссылка для крона';
+$_['entry_name']                  = 'Название блока';
+$_['entry_title']                 = 'Заголовок блока';
+$_['entry_type']                  = 'Тип блока';
 $_['entry_limit']                 = 'Лимит';
 $_['entry_width']                 = 'Ширина изображения';
 $_['entry_height']                = 'Высота изображения';
@@ -89,6 +77,7 @@ $_['entry_css']                   = 'Загружать CSS-стили из ка
 $_['entry_discount']              = 'Только акционные товары';
 $_['entry_brands']                = 'Товары только этих брендов';
 $_['entry_exclude_brands']        = 'Исключить товары этих брендов';
+$_['entry_block_status']          = 'Статус блока';
 
 // Error
 $_['error_permission']            = 'У вас нет прав для изменения модуля REES46!';